requirements.txt
azure-ai-documentintelligence
azure-cosmos
openai
python-dotenv
pydantic
